refactor: InstrumentSpec uses Decimal for precision

+++ barter-instrument/src/instrument/spec.rs
@@ -1,3 +1,4 @@
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, PartialEq, PartialOrd, Deserialize, Serialize)]
@@ -9,15 +10,15 @@ pub struct InstrumentSpec<AssetKey> {
 
 #[derive(Debug, Clone, PartialEq, PartialOrd, Deserialize, Serialize)]
 pub struct InstrumentSpecPrice {
-    pub min: f64,
-    pub tick_size: f64,
+    pub min: Decimal,
+    pub tick_size: Decimal,
 }
 
 #[derive(Debug, Clone, PartialEq, PartialOrd, Deserialize, Serialize)]
 pub struct InstrumentSpecQuantity<AssetKey> {
     pub unit: OrderQuantityUnits<AssetKey>,
-    pub min: f64,
-    pub increment: f64,
+    pub min: Decimal,
+    pub increment: Decimal,
 }
 
 #[derive(Debug, Clone, Copy, Eq, PartialEq, Ord, PartialOrd, Hash, Deserialize, Serialize)]
@@ -29,5 +30,5 @@ pub enum OrderQuantityUnits<AssetKey> {
 
 #[derive(Debug, Clone, PartialEq, PartialOrd, Deserialize, Serialize)]
 pub struct InstrumentSpecNotional {
-    pub min: f64,
+    pub min: Decimal,
 }
